+++ internal/handlers/status.go
@@ -1,0 +1,45 @@
+package handlers
+
+import (
+	"fmt"
+	"net/http"
+	"strings"
+
+	"github.com/containeroo/heartbeats/internal"
+	"github.com/gorilla/mux"
+	log "github.com/sirupsen/logrus"
+)
+
+// Status is the handler for the /status endpoint
+func Status(w http.ResponseWriter, req *http.Request) {
+	log.Tracef("%s %s%s", req.Method, req.RequestURI, strings.TrimSpace(req.URL.RawQuery))
+
+	outputFormat := GetOutputFormat(req)
+
+	vars := mux.Vars(req)
+	heartbeatName := vars["heartbeat"]
+
+	var txtFormat string = `Name: {{ .Name }}
+Status: {{ if .Status }}{{ .Status }}{{ else }}-{{ end }}
+LastPing: {{ .TimeAgo .LastPing }}`
+
+	if heartbeatName == "" {
+		textTmpl := fmt.Sprintf("%s%s\n%s", "{{ range . }}", txtFormat, "{{end}}")
+		WriteOutput(w, http.StatusOK, outputFormat, &internal.HeartbeatsServer.Heartbeats, textTmpl)
+		return
+	}
+
+	heartbeat, err := internal.HeartbeatsServer.GetHeartbeatByName(heartbeatName)
+	if err != nil {
+		WriteOutput(w, http.StatusNotFound, outputFormat, ResponseStatus{Status: "nok", Error: err.Error()}, "Status: {{ .Status }}\nError: {{  .Error }}")
+		return
+	}
+
+	heartbeatStates := &HeartbeatStatus{
+		Name:     heartbeat.Name,
+		Status:   heartbeat.Status,
+		LastPing: heartbeat.LastPing,
+	}
+
+	WriteOutput(w, http.StatusOK, outputFormat, &heartbeatStates, txtFormat)
+}
