+++ internal/server/server.go
@@ -1,8 +1,7 @@
-package internal
+package server
 
 import (
 	"context"
-	"embed"
 	"fmt"
 	"io/fs"
 
@@ -12,6 +11,8 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/containeroo/heartbeats/internal"
+	"github.com/containeroo/heartbeats/internal/handlers"
 	"github.com/containeroo/heartbeats/internal/metrics"
 	"github.com/gorilla/mux"
 	"github.com/prometheus/client_golang/prometheus"
@@ -19,15 +20,6 @@ import (
 
 	log "github.com/sirupsen/logrus"
 )
-
-var StaticFs embed.FS
-
-// Server is the holds HTTP server settings
-type Server struct {
-	Hostname string `mapstructure:"hostname"`
-	Port     int    `mapstructure:"port"`
-	SiteRoot string `mapstructure:"siteRoot"`
-}
 
 // NewRouter generates the router used in the HTTP Server
 func NewRouter() *mux.Router {
@@ -38,7 +30,7 @@ func NewRouter() *mux.Router {
 	metrics.PromMetrics = *metrics.NewMetrics(reg)
 
 	// handler for embed static files
-	fsys := fs.FS(StaticFs)
+	fsys := fs.FS(internal.StaticFS)
 	contentStatic, _ := fs.Sub(fsys, "web/static")
 	fs := http.FileServer(http.FS(contentStatic))
 	s := http.StripPrefix("/static/", fs)
@@ -46,28 +38,29 @@ func NewRouter() *mux.Router {
 	http.Handle("/", router)
 
 	//register handlers
-	router.HandleFunc("/", HandlerHome)
-	router.HandleFunc("/config", HandlerConfig)
-	router.HandleFunc("/version", HandlerVersion)
-	router.HandleFunc("/healthz", HandlerHealthz).Methods("GET", "POST")
-	router.HandleFunc("/ping", HandlerPingHelp).Methods("GET", "POST").Methods("GET", "POST")
-	router.HandleFunc("/ping/{heartbeat:[a-zA-Z0-9 _-]+}/fail", HandlerPingFail).Methods("GET", "POST")
-	router.HandleFunc("/ping/{heartbeat:[a-zA-Z0-9 _-]+}", HandlerPing).Methods("GET", "POST")
-	router.HandleFunc("/history", HandlerHistory)
-	router.HandleFunc("/history/{heartbeat:[a-zA-Z0-9 _-]+}", HandlerHistory)
-	router.HandleFunc("/status", HandlerStatus).Methods("GET", "POST")
-	router.HandleFunc("/status/{heartbeat:[a-zA-Z0-9 _-]+}", HandlerStatus).Methods("GET", "POST")
-	router.HandleFunc("/dashboard", HandlerDashboard)
-	router.HandleFunc("/docs", HandlerDocs)
-	router.HandleFunc("/docs/{chapter:[a-zA-Z0-9 _-]+}", HandlerChapter)
+	router.HandleFunc("/", handlers.Home)
+	router.HandleFunc("/config", handlers.Config)
+	router.HandleFunc("/version", handlers.Version)
+	router.HandleFunc("/healthz", handlers.Healthz).Methods("GET", "POST")
+	router.HandleFunc("/ping", handlers.PingHelp).Methods("GET", "POST").Methods("GET", "POST")
+	router.HandleFunc("/ping/{heartbeat:[a-zA-Z0-9 _-]+}/fail", handlers.PingFail).Methods("GET", "POST")
+	router.HandleFunc("/ping/{heartbeat:[a-zA-Z0-9 _-]+}", handlers.Ping).Methods("GET", "POST")
+	router.HandleFunc("/history", handlers.History)
+	router.HandleFunc("/history/{heartbeat:[a-zA-Z0-9 _-]+}", handlers.History)
+	router.HandleFunc("/status", handlers.Status).Methods("GET", "POST")
+	router.HandleFunc("/status/{heartbeat:[a-zA-Z0-9 _-]+}", handlers.Status).Methods("GET", "POST")
+	router.HandleFunc("/dashboard", handlers.Dashboard)
+	router.HandleFunc("/docs", handlers.Docs)
+	router.HandleFunc("/docs/{chapter:[a-zA-Z0-9 _-]+}", handlers.Chapter)
 	router.Handle("/metrics", promhttp.HandlerFor(reg, promhttp.HandlerOpts{Registry: reg})).Methods("GET", "POST")
-	router.NotFoundHandler = http.HandlerFunc(HandlerNotFound)
+	router.NotFoundHandler = http.HandlerFunc(handlers.NotFound)
 
 	return router
 }
 
 // Run will run the HTTP Server
-func (h *Heartbeats) Run() {
+func RunServer(hostname string, port int) {
+
 	// Set up a channel to listen to for interrupt signals
 	var runChan = make(chan os.Signal, 1)
 
@@ -81,7 +74,7 @@ func (h *Heartbeats) Run() {
 
 	// Define server options
 	server := &http.Server{
-		Addr:         fmt.Sprintf("%s:%d", h.Server.Hostname, h.Server.Port),
+		Addr:         fmt.Sprintf("%s:%d", hostname, port),
 		Handler:      NewRouter(),
 		ReadTimeout:  15 * time.Second,
 		WriteTimeout: 10 * time.Second,
